chore(alembic): squash migration history to fresh postgres baseline
