--- backend/alembic/versions/1caf6ee11f3c_m3_skills_marketplace.py
@@ -1,114 +1,0 @@
-"""m3_skills_marketplace
-
-Revision ID: 1caf6ee11f3c
-Revises: 63d571d8d2ac
-Create Date: 2026-04-27 11:19:03.111022
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel  # noqa: F401
-
-
-# revision identifiers, used by Alembic.
-revision: str = '1caf6ee11f3c'
-down_revision: Union[str, Sequence[str], None] = '63d571d8d2ac'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('org_preinstalled_tombstones',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('hidden_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('hidden_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('org_id', 'skill_id', name='uq_org_preinstalled_tombstone')
-    )
-    op.create_index(op.f('ix_org_preinstalled_tombstones_org_id'), 'org_preinstalled_tombstones', ['org_id'], unique=False)
-    op.create_index(op.f('ix_org_preinstalled_tombstones_skill_id'), 'org_preinstalled_tombstones', ['skill_id'], unique=False)
-    op.create_table('org_skill_installs',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('installed_version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('installed_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('installed_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('org_id', 'skill_id', name='uq_org_skill_install')
-    )
-    op.create_index(op.f('ix_org_skill_installs_org_id'), 'org_skill_installs', ['org_id'], unique=False)
-    op.create_index(op.f('ix_org_skill_installs_skill_id'), 'org_skill_installs', ['skill_id'], unique=False)
-    op.create_table('skill_versions',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
-    sa.Column('keywords', sa.JSON(), nullable=True),
-    sa.Column('raw_metadata', sa.JSON(), nullable=True),
-    sa.Column('storage_prefix', sqlmodel.sql.sqltypes.AutoString(length=512), nullable=False),
-    sa.Column('entry_file', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
-    sa.Column('uploaded_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('skill_id', 'version', name='uq_skill_version')
-    )
-    op.create_index(op.f('ix_skill_versions_skill_id'), 'skill_versions', ['skill_id'], unique=False)
-    op.create_table('skills',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
-    sa.Column('source', sqlmodel.sql.sqltypes.AutoString(length=16), nullable=False),
-    sa.Column('owner_org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=True),
-    sa.Column('current_version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
-    sa.Column('keywords', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name', name='uq_skill_name')
-    )
-    op.create_index('ix_skill_source_owner', 'skills', ['source', 'owner_org_id'], unique=False)
-    op.create_index(op.f('ix_skills_owner_org_id'), 'skills', ['owner_org_id'], unique=False)
-    op.create_table('workspace_skill_bindings',
-    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('org_skill_install_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
-    sa.Column('enabled', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('workspace_id', 'org_skill_install_id', name='uq_workspace_skill_binding')
-    )
-    op.create_index(op.f('ix_workspace_skill_bindings_org_id'), 'workspace_skill_bindings', ['org_id'], unique=False)
-    op.create_index(op.f('ix_workspace_skill_bindings_org_skill_install_id'), 'workspace_skill_bindings', ['org_skill_install_id'], unique=False)
-    op.create_index(op.f('ix_workspace_skill_bindings_workspace_id'), 'workspace_skill_bindings', ['workspace_id'], unique=False)
-    op.create_index('ix_wsb_org_ws', 'workspace_skill_bindings', ['org_id', 'workspace_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_wsb_org_ws', table_name='workspace_skill_bindings')
-    op.drop_index(op.f('ix_workspace_skill_bindings_workspace_id'), table_name='workspace_skill_bindings')
-    op.drop_index(op.f('ix_workspace_skill_bindings_org_skill_install_id'), table_name='workspace_skill_bindings')
-    op.drop_index(op.f('ix_workspace_skill_bindings_org_id'), table_name='workspace_skill_bindings')
-    op.drop_table('workspace_skill_bindings')
-    op.drop_index(op.f('ix_skills_owner_org_id'), table_name='skills')
-    op.drop_index('ix_skill_source_owner', table_name='skills')
-    op.drop_table('skills')
-    op.drop_index(op.f('ix_skill_versions_skill_id'), table_name='skill_versions')
-    op.drop_table('skill_versions')
-    op.drop_index(op.f('ix_org_skill_installs_skill_id'), table_name='org_skill_installs')
-    op.drop_index(op.f('ix_org_skill_installs_org_id'), table_name='org_skill_installs')
-    op.drop_table('org_skill_installs')
-    op.drop_index(op.f('ix_org_preinstalled_tombstones_skill_id'), table_name='org_preinstalled_tombstones')
-    op.drop_index(op.f('ix_org_preinstalled_tombstones_org_id'), table_name='org_preinstalled_tombstones')
-    op.drop_table('org_preinstalled_tombstones')
-    # ### end Alembic commands ###
