--- backend/alembic/versions/1d1dab71f0fa_widen_identity_id_columns_to_36_chars_.py
@@ -1,184 +1,0 @@
-"""widen identity id columns to 36 chars for uuid7
-
-Revision ID: 1d1dab71f0fa
-Revises: d3a4dffee1e1
-Create Date: 2026-04-17 16:49:41.037531
-
-"""
-from typing import Sequence, Union
-
-import sqlmodel  # noqa: F401
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '1d1dab71f0fa'
-down_revision: Union[str, Sequence[str], None] = 'd3a4dffee1e1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('agent_configs', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('agent_configs', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('agent_configs', 'id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('artifact_versions', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('artifact_versions', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('artifacts', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('artifacts', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('conversations', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('conversations', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('invite_tokens', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('invite_tokens', 'created_by',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('memberships', 'user_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('memberships', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('organizations', 'id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('user_sandboxes', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('user_sandboxes', 'workspace_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('users', 'id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('workspaces', 'id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    op.alter_column('workspaces', 'org_id',
-               existing_type=mysql.VARCHAR(length=32),
-               type_=sqlmodel.sql.sqltypes.AutoString(length=36),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('workspaces', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('workspaces', 'id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('users', 'id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('user_sandboxes', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('user_sandboxes', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('organizations', 'id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('memberships', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('memberships', 'user_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('invite_tokens', 'created_by',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('invite_tokens', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('conversations', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('conversations', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('artifacts', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('artifacts', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('artifact_versions', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('artifact_versions', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('agent_configs', 'id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('agent_configs', 'workspace_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    op.alter_column('agent_configs', 'org_id',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(length=36),
-               type_=mysql.VARCHAR(length=32),
-               existing_nullable=False)
-    # ### end Alembic commands ###
