--- backend/alembic/versions/3c453a5e6ad5_add_skill_deprecated_at.py
@@ -1,32 +1,0 @@
-"""add skill deprecated_at
-
-Revision ID: 3c453a5e6ad5
-Revises: 1caf6ee11f3c
-Create Date: 2026-04-27 17:46:23.825214
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3c453a5e6ad5'
-down_revision: Union[str, Sequence[str], None] = '1caf6ee11f3c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('skills', sa.Column('deprecated_at', sa.DateTime(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('skills', 'deprecated_at')
-    # ### end Alembic commands ###
