--- backend/alembic/versions/704767f28f2b_create_attachments_table.py
@@ -1,57 +1,0 @@
-"""create attachments table
-
-Revision ID: 704767f28f2b
-Revises: ed71c37a944d
-Create Date: 2026-04-28 12:50:00.000000
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlmodel import SQLModel
-
-
-# revision identifiers, used by Alembic.
-revision: str = '704767f28f2b'
-down_revision: Union[str, Sequence[str], None] = '9473174c2ef3'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'attachments',
-        sa.Column('id', sa.String(length=36), nullable=False),
-        sa.Column('org_id', sa.String(length=36), nullable=False),
-        sa.Column('workspace_id', sa.String(length=36), nullable=False),
-        sa.Column('conversation_id', sa.String(length=36), nullable=False),
-        sa.Column('uploader_user_id', sa.String(length=36), nullable=False),
-        sa.Column('filename', sa.String(length=255), nullable=False),
-        sa.Column('mime_type', sa.String(length=128), nullable=False),
-        sa.Column('size_bytes', sa.Integer(), nullable=False),
-        sa.Column('kind', sa.String(length=16), nullable=False),
-        sa.Column('object_key', sa.String(length=1024), nullable=False),
-        sa.Column('sandbox_path', sa.String(length=1024), nullable=False),
-        sa.Column('thumbnail_object_key', sa.String(length=1024), nullable=True),
-        sa.Column('width', sa.Integer(), nullable=True),
-        sa.Column('height', sa.Integer(), nullable=True),
-        sa.Column('status', sa.String(length=16), nullable=False, server_default='pending'),
-        sa.Column('attached_at', sa.DateTime(), nullable=True),
-        sa.Column('created_at', sa.DateTime(), nullable=False),
-        sa.Column('updated_at', sa.DateTime(), nullable=False),
-        sa.ForeignKeyConstraint(['conversation_id'], ['conversations.id'], ),
-        sa.PrimaryKeyConstraint('id'),
-        sa.Index('ix_attachments_conv_status', 'conversation_id', 'status'),
-        sa.Index('ix_attachments_org_ws', 'org_id', 'workspace_id'),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('attachments')
-    # ### end Alembic commands ###
