+++ backend/alembic/versions/7bb73b55aa6a_initial_postgres_schema.py
@@ -1,0 +1,370 @@
+"""initial postgres schema
+
+Revision ID: 7bb73b55aa6a
+Revises: 
+Create Date: 2026-04-29 23:43:04.809702
+
+"""
+from typing import Sequence, Union
+
+import sqlmodel  # noqa: F401
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '7bb73b55aa6a'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('agent_configs',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('system_prompt', sa.Text(), nullable=False),
+    sa.Column('model_id', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
+    sa.Column('skill_ids', sa.JSON(), nullable=True),
+    sa.Column('mcp_server_ids', sa.JSON(), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('workspace_id', name='uk_agent_configs_workspace')
+    )
+    op.create_index(op.f('ix_agent_configs_org_id'), 'agent_configs', ['org_id'], unique=False)
+    op.create_index('ix_agent_configs_org_ws', 'agent_configs', ['org_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_agent_configs_workspace_id'), 'agent_configs', ['workspace_id'], unique=False)
+    op.create_table('billing_events',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('conversation_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('event_type', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('cost_amount_micro', sa.Integer(), nullable=False),
+    sa.Column('currency', sqlmodel.sql.sqltypes.AutoString(length=3), nullable=False),
+    sa.Column('started_at', sa.DateTime(), nullable=False),
+    sa.Column('ended_at', sa.DateTime(), nullable=False),
+    sa.Column('duration_ms', sa.Integer(), nullable=False),
+    sa.Column('status', sqlmodel.sql.sqltypes.AutoString(length=20), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('ix_billing_events_conversation', 'billing_events', ['conversation_id'], unique=False)
+    op.create_index(op.f('ix_billing_events_org_id'), 'billing_events', ['org_id'], unique=False)
+    op.create_index('ix_billing_events_org_time', 'billing_events', ['org_id', 'started_at'], unique=False)
+    op.create_index('ix_billing_events_org_ws_time', 'billing_events', ['org_id', 'workspace_id', 'started_at'], unique=False)
+    op.create_index(op.f('ix_billing_events_user_id'), 'billing_events', ['user_id'], unique=False)
+    op.create_index(op.f('ix_billing_events_workspace_id'), 'billing_events', ['workspace_id'], unique=False)
+    op.create_table('conversations',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('creator_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('title', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_conversations_org_id'), 'conversations', ['org_id'], unique=False)
+    op.create_index('ix_conversations_user_ws', 'conversations', ['creator_user_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_conversations_workspace_id'), 'conversations', ['workspace_id'], unique=False)
+    op.create_table('invite_tokens',
+    sa.Column('token', sqlmodel.sql.sqltypes.AutoString(length=64), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('role', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('created_by', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('expires_at', sa.DateTime(), nullable=False),
+    sa.Column('used_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('token')
+    )
+    op.create_index('ix_invite_tokens_expires', 'invite_tokens', ['expires_at'], unique=False)
+    op.create_table('memberships',
+    sa.Column('user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('role', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('user_id', 'workspace_id')
+    )
+    op.create_table('org_preinstalled_tombstones',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('hidden_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('hidden_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('org_id', 'skill_id', name='uq_org_preinstalled_tombstone')
+    )
+    op.create_index(op.f('ix_org_preinstalled_tombstones_org_id'), 'org_preinstalled_tombstones', ['org_id'], unique=False)
+    op.create_index(op.f('ix_org_preinstalled_tombstones_skill_id'), 'org_preinstalled_tombstones', ['skill_id'], unique=False)
+    op.create_table('org_skill_installs',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('installed_version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('installed_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('installed_at', sa.DateTime(), nullable=False),
+    sa.Column('auto_bind', sa.Boolean(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('org_id', 'skill_id', name='uq_org_skill_install')
+    )
+    op.create_index(op.f('ix_org_skill_installs_org_id'), 'org_skill_installs', ['org_id'], unique=False)
+    op.create_index(op.f('ix_org_skill_installs_skill_id'), 'org_skill_installs', ['skill_id'], unique=False)
+    op.create_table('organizations',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('slug', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_organizations_slug'), 'organizations', ['slug'], unique=True)
+    op.create_table('skill_versions',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('skill_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('keywords', sa.JSON(), nullable=True),
+    sa.Column('raw_metadata', sa.JSON(), nullable=True),
+    sa.Column('storage_prefix', sqlmodel.sql.sqltypes.AutoString(length=512), nullable=False),
+    sa.Column('entry_file', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
+    sa.Column('uploaded_by_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('skill_id', 'version', name='uq_skill_version')
+    )
+    op.create_index(op.f('ix_skill_versions_skill_id'), 'skill_versions', ['skill_id'], unique=False)
+    op.create_table('skills',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
+    sa.Column('source', sqlmodel.sql.sqltypes.AutoString(length=16), nullable=False),
+    sa.Column('owner_org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=True),
+    sa.Column('current_version', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('keywords', sa.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.Column('deprecated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name', name='uq_skill_name')
+    )
+    op.create_index('ix_skill_source_owner', 'skills', ['source', 'owner_org_id'], unique=False)
+    op.create_index(op.f('ix_skills_owner_org_id'), 'skills', ['owner_org_id'], unique=False)
+    op.create_table('user_sandboxes',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('user_id', sqlmodel.sql.sqltypes.AutoString(length=64), nullable=False),
+    sa.Column('sandbox_id', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('status', sqlmodel.sql.sqltypes.AutoString(length=20), nullable=False),
+    sa.Column('image', sqlmodel.sql.sqltypes.AutoString(length=512), nullable=False),
+    sa.Column('volumes_config', sa.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('last_activity_at', sa.DateTime(), nullable=False),
+    sa.Column('ttl_seconds', sa.Integer(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('sandbox_id')
+    )
+    op.create_index(op.f('ix_user_sandboxes_org_id'), 'user_sandboxes', ['org_id'], unique=False)
+    op.create_index('ix_user_sandboxes_org_ws', 'user_sandboxes', ['org_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_user_sandboxes_user_id'), 'user_sandboxes', ['user_id'], unique=False)
+    op.create_index('ix_user_sandboxes_user_ws_status', 'user_sandboxes', ['user_id', 'workspace_id', 'status'], unique=False)
+    op.create_index(op.f('ix_user_sandboxes_workspace_id'), 'user_sandboxes', ['workspace_id'], unique=False)
+    op.create_table('users',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('email', sqlmodel.sql.sqltypes.AutoString(length=320), nullable=False),
+    sa.Column('hashed_password', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('is_superuser', sa.Boolean(), nullable=False),
+    sa.Column('is_verified', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('language', sqlmodel.sql.sqltypes.AutoString(length=10), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_table('workspace_skill_bindings',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('org_skill_install_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('enabled', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('workspace_id', 'org_skill_install_id', name='uq_workspace_skill_binding')
+    )
+    op.create_index(op.f('ix_workspace_skill_bindings_org_id'), 'workspace_skill_bindings', ['org_id'], unique=False)
+    op.create_index(op.f('ix_workspace_skill_bindings_org_skill_install_id'), 'workspace_skill_bindings', ['org_skill_install_id'], unique=False)
+    op.create_index(op.f('ix_workspace_skill_bindings_workspace_id'), 'workspace_skill_bindings', ['workspace_id'], unique=False)
+    op.create_index('ix_wsb_org_ws', 'workspace_skill_bindings', ['org_id', 'workspace_id'], unique=False)
+    op.create_table('workspaces',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('ix_workspaces_org', 'workspaces', ['org_id'], unique=False)
+    op.create_table('artifacts',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('conversation_id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('artifact_type', sqlmodel.sql.sqltypes.AutoString(length=50), nullable=False),
+    sa.Column('path', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('entry_file', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('mime_type', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=True),
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=True),
+    sa.Column('version', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['conversation_id'], ['conversations.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_artifacts_conversation_id'), 'artifacts', ['conversation_id'], unique=False)
+    op.create_index(op.f('ix_artifacts_org_id'), 'artifacts', ['org_id'], unique=False)
+    op.create_index('ix_artifacts_org_ws', 'artifacts', ['org_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_artifacts_workspace_id'), 'artifacts', ['workspace_id'], unique=False)
+    op.create_table('attachments',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('conversation_id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('uploader_user_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('filename', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('mime_type', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
+    sa.Column('size_bytes', sa.Integer(), nullable=False),
+    sa.Column('kind', sqlmodel.sql.sqltypes.AutoString(length=16), nullable=False),
+    sa.Column('object_key', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('sandbox_path', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('thumbnail_object_key', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=True),
+    sa.Column('width', sa.Integer(), nullable=True),
+    sa.Column('height', sa.Integer(), nullable=True),
+    sa.Column('status', sqlmodel.sql.sqltypes.AutoString(length=16), nullable=False),
+    sa.Column('attached_at', sa.DateTime(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['conversation_id'], ['conversations.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('ix_attachments_conv_status', 'attachments', ['conversation_id', 'status'], unique=False)
+    op.create_index(op.f('ix_attachments_conversation_id'), 'attachments', ['conversation_id'], unique=False)
+    op.create_index(op.f('ix_attachments_org_id'), 'attachments', ['org_id'], unique=False)
+    op.create_index('ix_attachments_org_ws', 'attachments', ['org_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_attachments_workspace_id'), 'attachments', ['workspace_id'], unique=False)
+    op.create_table('billing_llm_events',
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('billing_event_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('provider', sqlmodel.sql.sqltypes.AutoString(length=64), nullable=False),
+    sa.Column('model_id', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
+    sa.Column('input_tokens', sa.Integer(), nullable=False),
+    sa.Column('output_tokens', sa.Integer(), nullable=False),
+    sa.Column('cache_read_tokens', sa.Integer(), nullable=False),
+    sa.Column('cache_write_tokens', sa.Integer(), nullable=False),
+    sa.Column('price_input_per_mtok_micro', sa.Integer(), nullable=False),
+    sa.Column('price_output_per_mtok_micro', sa.Integer(), nullable=False),
+    sa.Column('price_cache_read_per_mtok_micro', sa.Integer(), nullable=False),
+    sa.Column('price_cache_write_per_mtok_micro', sa.Integer(), nullable=False),
+    sa.Column('parent_run_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=True),
+    sa.Column('subagent_depth', sa.Integer(), nullable=False),
+    sa.Column('error_class', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=True),
+    sa.ForeignKeyConstraint(['billing_event_id'], ['billing_events.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_billing_llm_events_billing_event_id'), 'billing_llm_events', ['billing_event_id'], unique=False)
+    op.create_index('ix_billing_llm_events_parent_run_id', 'billing_llm_events', ['parent_run_id'], unique=False)
+    op.create_index('ix_billing_llm_events_provider_model', 'billing_llm_events', ['provider', 'model_id'], unique=False)
+    op.create_table('artifact_versions',
+    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=36), nullable=False),
+    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('artifact_id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('version', sa.Integer(), nullable=False),
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=True),
+    sa.Column('path', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
+    sa.Column('entry_file', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('mime_type', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['artifact_id'], ['artifacts.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_artifact_versions_artifact_id'), 'artifact_versions', ['artifact_id'], unique=False)
+    op.create_index(op.f('ix_artifact_versions_org_id'), 'artifact_versions', ['org_id'], unique=False)
+    op.create_index('ix_artifact_versions_org_ws', 'artifact_versions', ['org_id', 'workspace_id'], unique=False)
+    op.create_index(op.f('ix_artifact_versions_workspace_id'), 'artifact_versions', ['workspace_id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_artifact_versions_workspace_id'), table_name='artifact_versions')
+    op.drop_index('ix_artifact_versions_org_ws', table_name='artifact_versions')
+    op.drop_index(op.f('ix_artifact_versions_org_id'), table_name='artifact_versions')
+    op.drop_index(op.f('ix_artifact_versions_artifact_id'), table_name='artifact_versions')
+    op.drop_table('artifact_versions')
+    op.drop_index('ix_billing_llm_events_provider_model', table_name='billing_llm_events')
+    op.drop_index('ix_billing_llm_events_parent_run_id', table_name='billing_llm_events')
+    op.drop_index(op.f('ix_billing_llm_events_billing_event_id'), table_name='billing_llm_events')
+    op.drop_table('billing_llm_events')
+    op.drop_index(op.f('ix_attachments_workspace_id'), table_name='attachments')
+    op.drop_index('ix_attachments_org_ws', table_name='attachments')
+    op.drop_index(op.f('ix_attachments_org_id'), table_name='attachments')
+    op.drop_index(op.f('ix_attachments_conversation_id'), table_name='attachments')
+    op.drop_index('ix_attachments_conv_status', table_name='attachments')
+    op.drop_table('attachments')
+    op.drop_index(op.f('ix_artifacts_workspace_id'), table_name='artifacts')
+    op.drop_index('ix_artifacts_org_ws', table_name='artifacts')
+    op.drop_index(op.f('ix_artifacts_org_id'), table_name='artifacts')
+    op.drop_index(op.f('ix_artifacts_conversation_id'), table_name='artifacts')
+    op.drop_table('artifacts')
+    op.drop_index('ix_workspaces_org', table_name='workspaces')
+    op.drop_table('workspaces')
+    op.drop_index('ix_wsb_org_ws', table_name='workspace_skill_bindings')
+    op.drop_index(op.f('ix_workspace_skill_bindings_workspace_id'), table_name='workspace_skill_bindings')
+    op.drop_index(op.f('ix_workspace_skill_bindings_org_skill_install_id'), table_name='workspace_skill_bindings')
+    op.drop_index(op.f('ix_workspace_skill_bindings_org_id'), table_name='workspace_skill_bindings')
+    op.drop_table('workspace_skill_bindings')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.drop_table('users')
+    op.drop_index(op.f('ix_user_sandboxes_workspace_id'), table_name='user_sandboxes')
+    op.drop_index('ix_user_sandboxes_user_ws_status', table_name='user_sandboxes')
+    op.drop_index(op.f('ix_user_sandboxes_user_id'), table_name='user_sandboxes')
+    op.drop_index('ix_user_sandboxes_org_ws', table_name='user_sandboxes')
+    op.drop_index(op.f('ix_user_sandboxes_org_id'), table_name='user_sandboxes')
+    op.drop_table('user_sandboxes')
+    op.drop_index(op.f('ix_skills_owner_org_id'), table_name='skills')
+    op.drop_index('ix_skill_source_owner', table_name='skills')
+    op.drop_table('skills')
+    op.drop_index(op.f('ix_skill_versions_skill_id'), table_name='skill_versions')
+    op.drop_table('skill_versions')
+    op.drop_index(op.f('ix_organizations_slug'), table_name='organizations')
+    op.drop_table('organizations')
+    op.drop_index(op.f('ix_org_skill_installs_skill_id'), table_name='org_skill_installs')
+    op.drop_index(op.f('ix_org_skill_installs_org_id'), table_name='org_skill_installs')
+    op.drop_table('org_skill_installs')
+    op.drop_index(op.f('ix_org_preinstalled_tombstones_skill_id'), table_name='org_preinstalled_tombstones')
+    op.drop_index(op.f('ix_org_preinstalled_tombstones_org_id'), table_name='org_preinstalled_tombstones')
+    op.drop_table('org_preinstalled_tombstones')
+    op.drop_table('memberships')
+    op.drop_index('ix_invite_tokens_expires', table_name='invite_tokens')
+    op.drop_table('invite_tokens')
+    op.drop_index(op.f('ix_conversations_workspace_id'), table_name='conversations')
+    op.drop_index('ix_conversations_user_ws', table_name='conversations')
+    op.drop_index(op.f('ix_conversations_org_id'), table_name='conversations')
+    op.drop_table('conversations')
+    op.drop_index(op.f('ix_billing_events_workspace_id'), table_name='billing_events')
+    op.drop_index(op.f('ix_billing_events_user_id'), table_name='billing_events')
+    op.drop_index('ix_billing_events_org_ws_time', table_name='billing_events')
+    op.drop_index('ix_billing_events_org_time', table_name='billing_events')
+    op.drop_index(op.f('ix_billing_events_org_id'), table_name='billing_events')
+    op.drop_index('ix_billing_events_conversation', table_name='billing_events')
+    op.drop_table('billing_events')
+    op.drop_index(op.f('ix_agent_configs_workspace_id'), table_name='agent_configs')
+    op.drop_index('ix_agent_configs_org_ws', table_name='agent_configs')
+    op.drop_index(op.f('ix_agent_configs_org_id'), table_name='agent_configs')
+    op.drop_table('agent_configs')
+    # ### end Alembic commands ###
