--- backend/alembic/versions/88ca1a8071c4_drop_messages_table.py
@@ -1,39 +1,0 @@
-"""drop_messages_table
-
-Revision ID: 88ca1a8071c4
-Revises: a1b2c3d4e5f6
-Create Date: 2026-03-31 12:28:24.189692
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '88ca1a8071c4'
-down_revision: Union[str, Sequence[str], None] = 'a1b2c3d4e5f6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    op.drop_table("messages")
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    op.create_table(
-        "messages",
-        sa.Column("id", sa.String(length=36), nullable=False),
-        sa.Column("conversation_id", sa.String(length=36), nullable=False),
-        sa.Column("role", sa.String(length=20), nullable=False),
-        sa.Column("content", sa.Text(), nullable=True),
-        sa.Column("events", sa.JSON(), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.ForeignKeyConstraint(["conversation_id"], ["conversations.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("ix_messages_conversation_id", "messages", ["conversation_id"])
