--- backend/alembic/versions/9473174c2ef3_add_user_language_field.py
@@ -1,33 +1,0 @@
-"""add user language field
-
-Revision ID: 9473174c2ef3
-Revises: ed71c37a944d
-Create Date: 2026-04-28 11:30:10.933845
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-
-
-# revision identifiers, used by Alembic.
-revision: str = '9473174c2ef3'
-down_revision: Union[str, Sequence[str], None] = 'ed71c37a944d'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('language', sqlmodel.sql.sqltypes.AutoString(length=10), nullable=False, server_default='en'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'language')
-    # ### end Alembic commands ###
