--- backend/alembic/versions/b744f6318cd3_create_conversations_messages.py
@@ -1,52 +1,0 @@
-"""create_conversations_messages
-
-Revision ID: b744f6318cd3
-Revises:
-Create Date: 2026-03-17 10:29:50.824453
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b744f6318cd3'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('conversations',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('title', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('messages',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('conversation_id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('role', sqlmodel.sql.sqltypes.AutoString(length=20), nullable=False),
-    sa.Column('content', sa.Text(), nullable=True),
-    sa.Column('events', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['conversation_id'], ['conversations.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_messages_conversation_id'), 'messages', ['conversation_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_messages_conversation_id'), table_name='messages')
-    op.drop_table('messages')
-    op.drop_table('conversations')
-    # ### end Alembic commands ###
