--- backend/alembic/versions/d3a4dffee1e1_m1_identity_and_scoping.py
@@ -1,172 +1,0 @@
-"""m1_identity_and_scoping
-
-Revision ID: d3a4dffee1e1
-Revises: ffce48f23c78
-Create Date: 2026-04-17 15:33:44.038723
-
-"""
-from typing import Sequence, Union
-
-import sqlmodel  # noqa: F401
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd3a4dffee1e1'
-down_revision: Union[str, Sequence[str], None] = 'ffce48f23c78'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('agent_configs',
-    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('system_prompt', sa.Text(), nullable=False),
-    sa.Column('model_id', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=False),
-    sa.Column('skill_ids', sa.JSON(), nullable=True),
-    sa.Column('mcp_server_ids', sa.JSON(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('workspace_id', name='uk_agent_configs_workspace')
-    )
-    op.create_index(op.f('ix_agent_configs_org_id'), 'agent_configs', ['org_id'], unique=False)
-    op.create_index('ix_agent_configs_org_ws', 'agent_configs', ['org_id', 'workspace_id'], unique=False)
-    op.create_index(op.f('ix_agent_configs_workspace_id'), 'agent_configs', ['workspace_id'], unique=False)
-    op.create_table('invite_tokens',
-    sa.Column('token', sqlmodel.sql.sqltypes.AutoString(length=64), nullable=False),
-    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('role', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('created_by', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('expires_at', sa.DateTime(), nullable=False),
-    sa.Column('used_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('token')
-    )
-    op.create_index('ix_invite_tokens_expires', 'invite_tokens', ['expires_at'], unique=False)
-    op.create_table('memberships',
-    sa.Column('user_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('role', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('user_id', 'workspace_id')
-    )
-    op.create_table('organizations',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('users',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('email', sqlmodel.sql.sqltypes.AutoString(length=320), nullable=False),
-    sa.Column('hashed_password', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=False),
-    sa.Column('is_superuser', sa.Boolean(), nullable=False),
-    sa.Column('is_verified', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_table('workspaces',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=False),
-    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index('ix_workspaces_org', 'workspaces', ['org_id'], unique=False)
-    op.add_column('artifact_versions', sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.add_column('artifact_versions', sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.create_index(op.f('ix_artifact_versions_org_id'), 'artifact_versions', ['org_id'], unique=False)
-    op.create_index('ix_artifact_versions_org_ws', 'artifact_versions', ['org_id', 'workspace_id'], unique=False)
-    op.create_index(op.f('ix_artifact_versions_workspace_id'), 'artifact_versions', ['workspace_id'], unique=False)
-    op.add_column('artifacts', sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.add_column('artifacts', sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.create_index(op.f('ix_artifacts_org_id'), 'artifacts', ['org_id'], unique=False)
-    op.create_index('ix_artifacts_org_ws', 'artifacts', ['org_id', 'workspace_id'], unique=False)
-    op.create_index(op.f('ix_artifacts_workspace_id'), 'artifacts', ['workspace_id'], unique=False)
-    op.add_column('conversations', sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.add_column('conversations', sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.create_index(op.f('ix_conversations_org_id'), 'conversations', ['org_id'], unique=False)
-    op.create_index('ix_conversations_org_ws', 'conversations', ['org_id', 'workspace_id'], unique=False)
-    op.create_index(op.f('ix_conversations_workspace_id'), 'conversations', ['workspace_id'], unique=False)
-    op.add_column('user_sandboxes', sa.Column('org_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.add_column('user_sandboxes', sa.Column('workspace_id', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True))
-    op.drop_index(op.f('ix_user_sandboxes_user_status'), table_name='user_sandboxes')
-    op.create_index(op.f('ix_user_sandboxes_org_id'), 'user_sandboxes', ['org_id'], unique=False)
-    op.create_index('ix_user_sandboxes_org_ws', 'user_sandboxes', ['org_id', 'workspace_id'], unique=False)
-    op.create_index('ix_user_sandboxes_user_ws_status', 'user_sandboxes', ['user_id', 'workspace_id', 'status'], unique=False)
-    op.create_index(op.f('ix_user_sandboxes_workspace_id'), 'user_sandboxes', ['workspace_id'], unique=False)
-    # ### end Alembic commands ###
-
-    # ---- 数据迁移：默认 org/ws + 回填 ----
-    from datetime import UTC, datetime
-
-    default_org_id = "default-org"
-    default_ws_id = "default-ws"
-    now = datetime.now(UTC).strftime("%Y-%m-%d %H:%M:%S")
-
-    op.execute(
-        f"INSERT INTO organizations (id, name, created_at) "
-        f"VALUES ('{default_org_id}', 'Default', '{now}')"
-    )
-    op.execute(
-        f"INSERT INTO workspaces (id, org_id, name, created_at) "
-        f"VALUES ('{default_ws_id}', '{default_org_id}', 'Default Workspace', '{now}')"
-    )
-
-    # 回填现有表
-    for tbl in ("conversations", "artifacts", "artifact_versions", "user_sandboxes"):
-        op.execute(
-            f"UPDATE {tbl} SET org_id = '{default_org_id}', "
-            f"workspace_id = '{default_ws_id}' WHERE org_id IS NULL"
-        )
-
-    # 把 nullable 列改成 NOT NULL
-    for tbl in ("conversations", "artifacts", "artifact_versions", "user_sandboxes"):
-        op.alter_column(tbl, "org_id", existing_type=sa.String(length=32), nullable=False)
-        op.alter_column(tbl, "workspace_id", existing_type=sa.String(length=32), nullable=False)
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_sandboxes_workspace_id'), table_name='user_sandboxes')
-    op.drop_index('ix_user_sandboxes_user_ws_status', table_name='user_sandboxes')
-    op.drop_index('ix_user_sandboxes_org_ws', table_name='user_sandboxes')
-    op.drop_index(op.f('ix_user_sandboxes_org_id'), table_name='user_sandboxes')
-    op.create_index(op.f('ix_user_sandboxes_user_status'), 'user_sandboxes', ['user_id', 'status'], unique=False)
-    op.drop_column('user_sandboxes', 'workspace_id')
-    op.drop_column('user_sandboxes', 'org_id')
-    op.drop_index(op.f('ix_conversations_workspace_id'), table_name='conversations')
-    op.drop_index('ix_conversations_org_ws', table_name='conversations')
-    op.drop_index(op.f('ix_conversations_org_id'), table_name='conversations')
-    op.drop_column('conversations', 'workspace_id')
-    op.drop_column('conversations', 'org_id')
-    op.drop_index(op.f('ix_artifacts_workspace_id'), table_name='artifacts')
-    op.drop_index('ix_artifacts_org_ws', table_name='artifacts')
-    op.drop_index(op.f('ix_artifacts_org_id'), table_name='artifacts')
-    op.drop_column('artifacts', 'workspace_id')
-    op.drop_column('artifacts', 'org_id')
-    op.drop_index(op.f('ix_artifact_versions_workspace_id'), table_name='artifact_versions')
-    op.drop_index('ix_artifact_versions_org_ws', table_name='artifact_versions')
-    op.drop_index(op.f('ix_artifact_versions_org_id'), table_name='artifact_versions')
-    op.drop_column('artifact_versions', 'workspace_id')
-    op.drop_column('artifact_versions', 'org_id')
-    op.drop_index('ix_workspaces_org', table_name='workspaces')
-    op.drop_table('workspaces')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    op.drop_table('organizations')
-    op.drop_table('memberships')
-    op.drop_index('ix_invite_tokens_expires', table_name='invite_tokens')
-    op.drop_table('invite_tokens')
-    op.drop_index(op.f('ix_agent_configs_workspace_id'), table_name='agent_configs')
-    op.drop_index('ix_agent_configs_org_ws', table_name='agent_configs')
-    op.drop_index(op.f('ix_agent_configs_org_id'), table_name='agent_configs')
-    op.drop_table('agent_configs')
-    # ### end Alembic commands ###
