--- backend/alembic/versions/ed71c37a944d_add_org_skill_install_auto_bind.py
@@ -1,32 +1,0 @@
-"""add_org_skill_install_auto_bind
-
-Revision ID: ed71c37a944d
-Revises: 3c453a5e6ad5
-Create Date: 2026-04-27 19:04:32.684706
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ed71c37a944d'
-down_revision: Union[str, Sequence[str], None] = '3c453a5e6ad5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('org_skill_installs', sa.Column('auto_bind', sa.Boolean(), server_default='0', nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('org_skill_installs', 'auto_bind')
-    # ### end Alembic commands ###
