--- backend/alembic/versions/ffce48f23c78_create_artifact_versions_table.py
@@ -1,46 +1,0 @@
-"""create artifact_versions table
-
-Revision ID: ffce48f23c78
-Revises: c7d8e9f0a1b2
-Create Date: 2026-04-09 09:51:17.857079
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-# revision identifiers, used by Alembic.
-revision: str = 'ffce48f23c78'
-down_revision: Union[str, Sequence[str], None] = 'c7d8e9f0a1b2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('artifact_versions',
-    sa.Column('id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('artifact_id', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('version', sa.Integer(), nullable=False),
-    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=True),
-    sa.Column('path', sqlmodel.sql.sqltypes.AutoString(length=1024), nullable=False),
-    sa.Column('entry_file', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
-    sa.Column('mime_type', sqlmodel.sql.sqltypes.AutoString(length=128), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['artifact_id'], ['artifacts.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_artifact_versions_artifact_id'), 'artifact_versions', ['artifact_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_artifact_versions_artifact_id'), table_name='artifact_versions')
-    op.drop_table('artifact_versions')
-    # ### end Alembic commands ###
